Dind successfully build on circleci using ubuntu:14.04
Using debian:jessie successfully allows me to build a Docker-in-Docker
image on my mac, however this failed on @circleci. The issue appears to
be somewhat related to modprobe though I am not clear how this could be
an issue when running inside an image, regardless of the host system.
Switching to ubuntu:14.04 allows this to build.

This is obviously a very obscure error but may be relevant at some
point in the future.

Signed-off-by: Michael Barton <mail@michaelbarton.me.uk>